+++ vm-device/Cargo.toml
@@ -14,4 +14,5 @@ hypervisor = { path = "../hypervisor" }
 serde = { workspace = true, features = ["derive", "rc"] }
 thiserror = { workspace = true }
 vfio-ioctls = { workspace = true, default-features = false }
+vm-memory = { workspace = true, features = ["backend-mmap"] }
 vmm-sys-util = { workspace = true }
